# add optional faster and more granular clock for getting current time