`udatetime.utcnow()` behavior:

## Linux

**new**:
- 113 ns ± 5.09 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
- includes useconds

**old**:
- 148 ns ± 2.22 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
- rounded to millis


## Mac
**new**:
- 93.2 ns ± 0.312 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
- includes useconds

**old**:
- 96.7 ns ± 0.718 ns per loop (mean ± std. dev. of 7 runs, 10,000,000 loops each)
- rounded to millis